backend/quality-service/requirements.txt
# AVA Image Quality Service - Python Dependencies

# Core image processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Flask API
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Additional utilities
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-cov==4.1.0